requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.14.0
openpyxl>=3.1.0
jupyter>=1.0.0
ipykernel>=6.25.0
scipy>=1.10.0
pvlib>=0.10.0
requests>=2.31.0
python-dotenv>=1.0.0

